requirements.txt
streamlit
pandas
pandas_ta
baostock
requests
python-dotenv
tqdm
plotly
